Refactored color menu out of ViewController
