+++ src/application/styles-resources.css
@@ -14,7 +14,7 @@
 	-fx-background-image: url('../resources/icons/placeholder-tool-icon-2-freepik.png');
 }
 
-#toolbar #color-selector {
+#toolbar #color-picker {
 	-fx-background-image: url('../resources/icons/colors-icon-cropped-freepik.png');
 }
